requirements.txt
streamlit
praat-parselmouth
pandas
numpy
scikit-learn